Add BillDetailDialog and integrate bill details view
Implemented BillDetailDialog with a dark-themed UI to display detailed bill information, including items and totals. Integrated the dialog into BillPage for viewing bill details. Also fixed unit price data type in ToolAndMaterialPage and updated dialog title in WorkerDetailDialog to French.

+++ src/constructpro/UI/ToolAndMaterialPage.java
@@ -124,7 +124,7 @@ public class ToolAndMaterialPage extends JPanel{
                     rs.getInt("itemId"),
                     rs.getString("name"),
                     rs.getString("quantity"),
-                    rs.getInt("unitPrice"),
+                    rs.getDouble("unitPrice"),
                     rs.getString("itemType"),
                     rs.getString("site_name"),
                     rs.getString("billDate")
@@ -160,7 +160,7 @@ public class ToolAndMaterialPage extends JPanel{
                     rs.getInt("itemId"),
                     rs.getString("name"),
                     rs.getString("quantity"),
-                    rs.getInt("unitPrice"),
+                    rs.getDouble("unitPrice"),
                     rs.getString("itemType"),
                     rs.getString("site_name"),
                     rs.getString("billDate")
